+++ mackerel-plugin-accesslog/lib/accesslog.go
@@ -2,6 +2,7 @@ package mpaccesslog
 
 import (
 	"bufio"
+	"bytes"
 	"flag"
 	"fmt"
 	"io"
@@ -24,6 +25,7 @@ type AccesslogPlugin struct {
 	prefix    string
 	file      string
 	posFile   string
+	parser    axslogparser.Parser
 	noPosFile bool
 }
 
@@ -117,18 +119,33 @@ func (p *AccesslogPlugin) FetchMetrics() (map[string]float64, error) {
 		ret[k] = 0
 	}
 	var reqtimes []float64
-	var psr axslogparser.Parser
-	s := bufio.NewScanner(rc)
-	for s.Scan() {
+	r := bufio.NewReader(rc)
+	for {
 		var (
 			l   *axslogparser.Log
 			err error
+			bb  bytes.Buffer
 		)
-		line := s.Text()
-		if psr == nil {
-			psr, l, err = axslogparser.GuessParser(line)
+		buf, isPrefix, err := r.ReadLine()
+		bb.Write(buf)
+		for isPrefix {
+			buf, isPrefix, err = r.ReadLine()
+			if err != nil {
+				break
+			}
+			bb.Write(buf)
+		}
+		if err != nil {
+			if err != io.EOF {
+				log.Println(err)
+			}
+			break
+		}
+		line := bb.String()
+		if p.parser == nil {
+			p.parser, l, err = axslogparser.GuessParser(line)
 		} else {
-			l, err = psr.Parse(line)
+			l, err = p.parser.Parse(line)
 		}
 		if err != nil {
 			log.Println(err)
@@ -142,9 +159,6 @@ func (p *AccesslogPlugin) FetchMetrics() (map[string]float64, error) {
 		} else if l.TakenSec != nil {
 			reqtimes = append(reqtimes, *l.TakenSec)
 		}
-	}
-	if s.Err() != nil {
-		log.Println(s.Err())
 	}
 	if ret["total_count"] > 0 {
 		for _, v := range []string{"2xx", "3xx", "4xx", "5xx"} {
@@ -169,6 +183,7 @@ func (p *AccesslogPlugin) FetchMetrics() (map[string]float64, error) {
 func Do() {
 	var (
 		optPrefix    = flag.String("metric-key-prefix", "", "Metric key prefix")
+		optFormat    = flag.String("format", "", "Access Log format ('ltsv' or 'apache')")
 		optPosFile   = flag.String("posfile", "", "(not necessary to specify it in the usual use case) posfile")
 		optNoPosFile = flag.Bool("no-posfile", false, "no position file")
 	)
@@ -181,10 +196,26 @@ func Do() {
 		flag.Usage()
 		os.Exit(1)
 	}
+
+	var parser axslogparser.Parser
+	switch *optFormat {
+	case "":
+		parser = nil // guess format by log (default)
+	case "ltsv":
+		parser = &axslogparser.LTSV{}
+	case "apache":
+		parser = &axslogparser.Apache{}
+	default:
+		fmt.Fprintf(os.Stderr, "Error: '%s' is invalid format name\n", *optFormat)
+		flag.Usage()
+		os.Exit(1)
+	}
+
 	mp.NewMackerelPlugin(&AccesslogPlugin{
 		prefix:    *optPrefix,
 		file:      flag.Args()[0],
 		posFile:   *optPosFile,
 		noPosFile: *optNoPosFile,
+		parser:    parser,
 	}).Run()
 }